+++ src/test/java/CucumberRunner.java
@@ -1,0 +1,3 @@
+public class CucumberRunner {
+
+}
