--- src/test/java/HomeWork24.java
@@ -1,13 +1,0 @@
-
-import org.testng.annotations.Test;
-import pages.LoginPage;
-@Test
-public class HomeWork24 extends BaseTest{
-    public void BrowserFactory(){
-        LoginPage loginpage = new LoginPage(getDriver());
-        loginpage.provideEmail("elzat.nurmanbetova@testpro.io");
-        loginpage.providePassword("kochkor123-Q!");
-        loginpage.clickSubmit();
-
-    }
-}
