+++ src/test/java/checkAllSongs.java
@@ -1,12 +1,13 @@
-import org.testng.Assert;
 import org.testng.annotations.Test;
 import pages.HomePage;
 import pages.LoginPage;
+import stepDefinitons.BaseDefinition;
+
 @Test
-public class HomeWork17 extends BaseTest{
+public class checkAllSongs extends BaseDefinition {
   public void viewAllSongs(){
-      LoginPage loginpage = new LoginPage(getDriver());
-      HomePage homepage = new HomePage(getDriver());
+      LoginPage loginpage = new LoginPage(driver);
+      HomePage homepage = new HomePage(driver);
       loginpage.provideEmail("elzat.nurmanbetova@testpro.io");
       loginpage.providePassword("kochkor123-Q!");
       loginpage.clickSubmit();
