+++ src/test/java/stepDefinitons/CucumberRunner.java
@@ -1,0 +1,31 @@
+package stepDefinitons;
+
+import io.cucumber.testng.AbstractTestNGCucumberTests;
+import io.cucumber.testng.CucumberOptions;
+import io.cucumber.testng.TestNGCucumberRunner;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.DataProvider;
+import org.testng.annotations.Test;
+
+
+    @CucumberOptions(features = {"src/test/resources/features"},
+            publish=true
+    )
+    @Test
+    public class CucumberRunner extends AbstractTestNGCucumberTests {
+        private TestNGCucumberRunner testNGCucumberRunner;
+        @BeforeClass(alwaysRun = true)
+        public void setUpCucumber(){
+            testNGCucumberRunner = new TestNGCucumberRunner(this.getClass());
+        }
+        @Override
+        @DataProvider(parallel = true)
+        public Object[][] scenarios(){
+            return super.scenarios();
+        }
+        @AfterClass(alwaysRun = true)
+        public void tearDownClass(){
+            testNGCucumberRunner.finish();
+        }
+    }
