+++ src/test/resources/CucumberRunner.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-<suite name="TestNG">
+<suite name="TestNG" parallel="methods" thread-count="4">
     <parameter name="url" value="https://qa.koel.app/"/>
-    <test name="HomeWork23" preserve-order="false">
+    <test name="CucumberRunner Test" preserve-order="false">
         <classes>
-            <class name="HomeWork23"/>
+            <class name="CucumberRunner"/>
+            <class name="playFirstSong"/>
         </classes>
     </test>
 </suite>
