--- src/test/resources/HomeWork18.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="Homework18">
-    <parameter name="baseUrl" value="https://qa.koel.app/#!/queue"/>
-    <test name="Homework18" preserve-order="false">
-        <classes>
-            <class name="HomeWork18">
-                <methods>
-                    <include name="playSong"/>
-                </methods>
-            </class>
-        </classes>
-    </test>
-</suite>
