--- src/test/resources/HomeWork22.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-<suite name="TestNG">
-    <parameter name="url" value="https://qa.koel.app/"/>
-
-    <test name="HomeWork22" preserve-order="false">
-        <classes>
-            <class name="HomeWork22"/>
-        </classes>
-    </test>
-</suite>
