--- src/test/resources/Homework17.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="Homework17">
-    <!-- Setting a parameter named "baseUrl" with the value "https://qa.koel.app". -->
-    <parameter name="baseUrl" value="https://qa.koel.app"/>
-    <test name="Example Test" preserve-order="false">
-        <!-- Including a class named "Homework17" in the test. -->
-        <classes>
-            <class name="HomeWork17"/>
-        </classes>
-    </test>
-</suite>
