+++ src/test/resources/SmokeTest.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+<suite name="TestNG" parallel="methods" thread-count="4">
+    <parameter name="url" value="https://qa.koel.app/"/>
+    <test name="CucumberRunner Test" preserve-order="false">
+        <classes>
+            <class name="validLoginDefinition"/>
+            <class name="checkAllSongs"/>
+            <class name="playFirstSong"/>
+        </classes>
+    </test>
+</suite>
