--- src/test/resources/TestNG.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-<suite name="TestNG">
-    <parameter name="url" value="https://qa.koel.app/"/>
-    <test name="TestNG" preserve-order="false">
-        <classes>
-            <class name="HomeWork27"/>
-        </classes>
-    </test>
-</suite>
