+++ src/test/resources/features/junit-platform.properties
@@ -1,0 +1,1 @@
+cucumber.publish.quiet = true
